Highly increased drawing performance
I noticed that if there were a lot of objects in a view the program would start lagging, a lot. The problem was that there was no checking if the object being drawn is inside a drawing constraint or not, thus objects were always being drawn (together with their children) no matter if they were actually visible on the screen or not. My little change just checks if the object being drawn is inside a constraint and if not then it just doesn't draw it, nor it's children. The FPS increase for me was from about 3-5 FPS to about 18-20 FPS, keeping in mind that, in ComputerCraft, 20 FPS is kind of the max.